Flush images so that they can load after conversion
This makes the image load more consistent even though it usually happens on its
own.

Without this, sometimes after a thumbnail is generated, emacs does not call
load-image on the image-spec again (probably using the cached version) so the
image stays blank. Calling flush allows the image to be redrawn from scratch.